requirements.txt
Django>=4.2.7,<4.3
Pillow>= 9.0, <10.1
psycopg2-binary
requests
django-extensions
django-filter
djangorestframework
djangorestframework-jsonapi
djangorestframework-simplejwt
gunicorn
whitenoise
PyJWT